backend/requirements.txt
alabaster==0.7.12
asgiref==3.5.2
Babel==2.10.3
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docutils==0.19
gunicorn==20.1.0
idna==3.3
imagesize==1.4.1
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.1.1
prototype==0.2
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
snowballstemmer==2.2.0
Sphinx==5.1.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stripe==3.5.0
urllib3==1.26.10
whitenoise==6.2.0
